Add in racks refactor suggestions from Chris
