+++ src/racks/racks-service.js
@@ -4,9 +4,14 @@ const RacksService = {
   getUserRacks(db, userId) {
     return db.from('ru_racks').select('*').where('user_id', userId);
   },
+  verifyRackExists(db, rackId) {
+    return db.from('ru_racks').count(1).where({ rack_id: rackId });
+  },
   getById(db, userId, rackId) {
-    return RacksService.getUserRacks(db, userId)
-      .where('rack_id', rackId)
+    return db
+      .from('ru_racks')
+      .select('*')
+      .where({ rack_id: rackId, user_id: userId })
       .first();
   },
   insertRack(db, userId, newRack) {
@@ -17,19 +22,19 @@ const RacksService = {
       .then(([rack]) => rack)
       .then((rack) => RacksService.getById(db, userId, rack.rack_id));
   },
-  updateRack(db, id, updatedRack) {
+  updateRack(db, userId, rackId, updatedRack) {
     return db
-      .select('*')
       .from('ru_racks')
-      .where('rack_id', id)
+      .select('*')
+      .where({ rack_id: rackId, user_id: userId })
       .first()
       .update(updatedRack);
   },
-  deleteRack(db, id) {
+  deleteRack(db, userId, rackId) {
     return db
-      .select('*')
       .from('ru_racks')
-      .where('rack_id', id)
+      .select('*')
+      .where({ rack_id: rackId, user_id: userId })
       .first()
       .delete();
   },
